+++ CHANGELOG.md
@@ -2,7 +2,9 @@ miru.ink / Changelog
 ====================
 
 ## Version 0.0.10 - Experimental
-- Update: package.json dependencies.
+- Add: New `default` slot on the `AvatarStd` component to replace `props.name` when no `props.avatar` is present.
+- Update: Change the `AvtaraGroup` limit class and label when not enough space is available.
+- Update: `package.json` dependencies.
 - Fix: TypeScript type issues on dts plugin / vite build process.
 
 ## Version 0.0.9 - Experimental